+++ core/modules/user.py
@@ -1,23 +1,52 @@
 import datetime
+import enum
+import functools
 import hashlib
 import hmac
 import json
 import logging
-import time
+import secrets
 import warnings
 from typing import Optional
-
 import user_agents
+import time
 from google.auth.transport import requests
 from google.oauth2 import id_token
 
-from viur.core import conf, db, email, errors, exposed, forceSSL, i18n, securitykey, session, skeleton, tasks, utils, \
-    current
+from viur.core import (
+    conf, current, db, email, errors, exposed, forceSSL, i18n,
+    securitykey, session, skeleton, tasks, utils
+)
 from viur.core.bones import *
-from viur.core.bones.password import pbkdf2
+from viur.core.bones.password import PBKDF2_DEFAULT_ITERATIONS, encode_password
 from viur.core.prototypes.list import List
 from viur.core.ratelimit import RateLimit
 from viur.core.securityheaders import extendCsp
+
+
+@functools.total_ordering
+class Status(enum.Enum):
+    """Status enum for a user
+
+    Has backwards compatibility to be comparable with non-enum values.
+    Will be removed with viur-core 4.0.0
+    """
+
+    UNSET = 0  # Status is unset
+    WAITING_FOR_EMAIL_VERIFICATION = 1  # Waiting for email verification
+    WAITING_FOR_ADMIN_VERIFICATION = 2  # Waiting for verification through admin
+    DISABLED = 5  # Account disabled
+    ACTIVE = 10  # Active
+
+    def __eq__(self, other):
+        if isinstance(other, Status):
+            return super().__eq__(other)
+        return self.value == other
+
+    def __lt__(self, other):
+        if isinstance(other, Status):
+            return super().__lt__(other)
+        return self.value < other
 
 
 class UserSkel(skeleton.Skeleton):
@@ -30,6 +59,16 @@ class UserSkel(skeleton.Skeleton):
         caseSensitive=False,
         searchable=True,
         unique=UniqueValue(UniqueLockMethod.SameValue, True, "Username already taken"),
+    )
+
+    firstname = StringBone(
+        descr="Firstname",
+        searchable=True,
+    )
+
+    lastname = StringBone(
+        descr="Lastname",
+        searchable=True,
     )
 
     # Properties required by custom auth
@@ -48,6 +87,16 @@ class UserSkel(skeleton.Skeleton):
         unique=UniqueValue(UniqueLockMethod.SameValue, False, "UID already in use"),
     )
 
+    sync = BooleanBone(
+        descr="Sync user data with OAuth-based services",
+        defaultValue=True,
+        params={
+            "tooltip":
+                "If set, user data like firstname and lastname is automatically kept synchronous with the information "
+                "stored at the OAuth service provider (e.g. Google Login)."
+        }
+    )
+
     gaeadmin = BooleanBone(
         descr="Is GAE Admin",
         defaultValue=False,
@@ -58,22 +107,16 @@ class UserSkel(skeleton.Skeleton):
     access = SelectBone(
         descr="Access rights",
         values=lambda: {
-
             right: i18n.translate("server.modules.user.accessright.%s" % right, defaultText=right)
-                for right in sorted(conf["viur.accessRights"])
+            for right in sorted(conf["viur.accessRights"])
         },
         multiple=True,
     )
 
     status = SelectBone(
         descr="Account status",
-        values={
-            1: "Waiting for email verification",
-            2: "Waiting for verification through admin",
-            5: "Account disabled",
-            10: "Active"
-        },
-        defaultValue=10,
+        values=Status,
+        defaultValue=Status.ACTIVE,
         required=True,
     )
 
@@ -193,45 +236,43 @@ class UserPassword:
         res = query.filter("name.idx >=", name).getEntry()
 
         if res is None:
-            res = {"password": {"pwhash": "-invalid-", "salt": "-invalid"}, "status": 0, "name": {}}
+            res = {"password": {"pwhash": "-invalid-", "salt": "-invalid-"}, "status": 0, "name": {}}
 
-        passwd = pbkdf2(password[:conf["viur.maxPasswordLength"]], (res.get("password", None) or {}).get("salt", ""))
-        isOkay = True
-
-        # We do this exactly that way to avoid timing attacks
+        password_data = res.get("password") or {}
+        # old password hashes used 1001 iterations
+        iterations = password_data.get("iterations", 1001)
+        passwd = encode_password(password, password_data.get("salt", ""), iterations)["pwhash"]
 
         # Check if the username matches
-        storedUserName = (res.get("name") or {}).get("idx", "")
-        if len(storedUserName) != len(name):
-            isOkay = False
-        else:
-            for x, y in zip(storedUserName, name):
-                if x != y:
-                    isOkay = False
+        stored_user_name = (res.get("name") or {}).get("idx", "")
+        is_okay = secrets.compare_digest(stored_user_name, name)
 
         # Check if the password matches
-        storedPasswordHash = (res.get("password", None) or {}).get("pwhash", "-invalid-")
-        if len(storedPasswordHash) != len(passwd):
-            isOkay = False
-        else:
-            for x, y in zip(storedPasswordHash, passwd):
-                if x != y:
-                    isOkay = False
+        stored_password_hash = password_data.get("pwhash", "-invalid-")
+        is_okay &= secrets.compare_digest(stored_password_hash, passwd)
 
         accountStatus: Optional[int] = None
         # Verify that this account isn't blocked
-        if res["status"] < 10:
-            if isOkay:
+        if res["status"] < Status.ACTIVE.value:
+            if is_okay:
                 # The username and password is valid, in this case we can inform that user about his account status
                 # (ie account locked or email verification pending)
                 accountStatus = res["status"]
-            isOkay = False
+            is_okay = False
 
-        if not isOkay:
+        if not is_okay:
             self.loginRateLimit.decrementQuota()  # Only failed login attempts will count to the quota
             skel = self.LoginSkel()
             return self.userModule.render.login(skel, loginFailed=True, accountStatus=accountStatus)
         else:
+            if iterations < PBKDF2_DEFAULT_ITERATIONS:
+                logging.info(f"Update password hash for user {name}.")
+                # re-hash the password with more iterations
+                skel = self.userModule.editSkel()
+                skel.setEntity(res)
+                skel["key"] = res.key
+                skel["password"] = password  # will be hashed on serialize
+                skel.toDB(clearUpdateTag=True)
             return self.userModule.continueAuthenticationFlow(self, res.key)
 
     @exposed
@@ -314,7 +355,7 @@ class UserPassword:
 
         # Update the password, save the user, reset his session and show the success-template
         user_skel["password"] = skel["password"]
-        user_skel.toDB(clearUpdateTag=False)
+        user_skel.toDB(update_relations=True)
 
         return self.userModule.render.view(None, tpl=self.passwordRecoverySuccessTemplate)
 
@@ -341,9 +382,9 @@ class UserPassword:
             data["userKey"].id_or_name):
             return self.userModule.render.view(None, tpl=self.verifyFailedTemplate)
         if self.registrationAdminVerificationRequired:
-            skel["status"] = 2
+            skel["status"] = Status.WAITING_FOR_ADMIN_VERIFICATION
         else:
-            skel["status"] = 10
+            skel["status"] = Status.ACTIVE
         skel.toDB()
         return self.userModule.render.view(skel, tpl=self.verifySuccessTemplate)
 
@@ -353,18 +394,18 @@ class UserPassword:
     def addSkel(self):
         """
             Prepare the add-Skel for rendering.
-            Currently only calls self.userModule.addSkel() and sets skel["status"].value depening on
+            Currently only calls self.userModule.addSkel() and sets skel["status"] depending on
             self.registrationEmailVerificationRequired and self.registrationAdminVerificationRequired
             :return: viur.core.skeleton.Skeleton
         """
         skel = self.userModule.addSkel()
 
         if self.registrationEmailVerificationRequired:
-            defaultStatusValue = 1
+            defaultStatusValue = Status.WAITING_FOR_EMAIL_VERIFICATION
         elif self.registrationAdminVerificationRequired:
-            defaultStatusValue = 2
+            defaultStatusValue = Status.WAITING_FOR_ADMIN_VERIFICATION
         else:  # No further verification required
-            defaultStatusValue = 10
+            defaultStatusValue = Status.ACTIVE
 
         skel.status.readOnly = True
         skel["status"] = defaultStatusValue
@@ -400,8 +441,8 @@ class UserPassword:
         if not securitykey.validate(skey, useSessionKey=True):
             raise errors.PreconditionFailed()
         skel.toDB()
-        if self.registrationEmailVerificationRequired and str(skel["status"]) == "1":
-            # The user will have to verify his email-address. Create an skey and send it to his address
+        if self.registrationEmailVerificationRequired and skel["status"] == Status.WAITING_FOR_EMAIL_VERIFICATION:
+            # The user will have to verify his email-address. Create a skey and send it to his address
             skey = securitykey.create(duration=60 * 60 * 24 * 7, userKey=utils.normalizeKey(skel["key"]),
                                       name=skel["name"])
             skel.skey = BaseBone(descr="Skey")
@@ -455,24 +496,40 @@ class GoogleAccount:
 
         # fixme: use self.userModule.baseSkel() for this later
         addSkel = skeleton.skeletonByKind(self.userModule.addSkel().kindName)  # Ensure that we have the full skeleton
-        userSkel = addSkel().all().filter("uid =", uid).getSkel()
-        if not userSkel:
+
+        update = False
+        if not (userSkel := addSkel().all().filter("uid =", uid).getSkel()):
             # We'll try again - checking if there's already an user with that email
-            userSkel = addSkel().all().filter("name.idx =", email.lower()).getSkel()
-            if not userSkel:  # Still no luck - it's a completely new user
+            if not (userSkel := addSkel().all().filter("name.idx =", email.lower()).getSkel()):
+                # Still no luck - it's a completely new user
                 if not self.registrationEnabled:
                     if userInfo.get("hd") and userInfo["hd"] in conf["viur.user.google.gsuiteDomains"]:
                         print("User is from domain - adding account")
                     else:
                         logging.warning("Denying registration of %s", email)
                         raise errors.Forbidden("Registration for new users is disabled")
+
                 userSkel = addSkel()  # We'll add a new user
+
             userSkel["uid"] = uid
             userSkel["name"] = email
-            isAdd = True
-        else:
-            isAdd = False
-        if isAdd:
+            update = True
+
+        # Take user information from Google, if wanted!
+        if userSkel["sync"]:
+
+            for target, source in {
+                "name": email,
+                "firstname": userInfo.get("given_name"),
+                "lastname": userInfo.get("family_name"),
+            }.items():
+
+                if userSkel[target] != source:
+                    userSkel[target] = source
+                    update = True
+
+        if update:
+            # TODO: Get access from IAM or similar
             # if users.is_current_user_admin():
             #    if not userSkel["access"]:
             #        userSkel["access"] = []
@@ -482,6 +539,7 @@ class GoogleAccount:
             # else:
             #    userSkel["gaeadmin"] = False
             assert userSkel.toDB()
+
         return self.userModule.continueAuthenticationFlow(self, userSkel["key"])
 
 
@@ -663,7 +721,7 @@ class User(List):
         user = current.user.get()
         if not (user and user["access"] and ("%s-add" % self.moduleName in user["access"] or "root" in user["access"])):
             skel.status.readOnly = True
-            skel["status"] = 0
+            skel["status"] = Status.UNSET
             skel.status.visible = False
             skel.access.readOnly = True
             skel["access"] = []
@@ -757,6 +815,10 @@ class User(List):
         skel = self.baseSkel()
         if not skel.fromDB(key):
             raise ValueError(f"Unable to authenticate unknown user {key}")
+
+        # Verify that this user account is active
+        if skel["status"] < Status.ACTIVE.value:
+            raise errors.Forbidden("The user is disabled and cannot be authenticated.")
 
         # Update session for user
         session = current.session.get()
@@ -867,12 +929,16 @@ class User(List):
 
         return json.dumps(res)
 
+    def onEdited(self, skel):
+        super().onEdited(skel)
+        # In case the user is set to inactive, kill all sessions
+        if skel["status"] < Status.ACTIVE.value:
+            session.killSessionByUser(skel["key"])
+
     def onDeleted(self, skel):
-        """
-            Invalidate all sessions of that user
-        """
-        super(User, self).onDeleted(skel)
-        session.killSessionByUser(str(skel["key"]))
+        super().onDeleted(skel)
+        # Invalidate all sessions of that user
+        session.killSessionByUser(skel["key"])
 
 
 @tasks.StartupTask
@@ -892,7 +958,7 @@ def createNewUserIfNotExists():
             uname = f"""admin@{conf["viur.instance.project_id"]}.appspot.com"""
             pw = utils.generateRandomString(13)
             addSkel["name"] = uname
-            addSkel["status"] = 10  # Ensure its enabled right away
+            addSkel["status"] = Status.ACTIVE  # Ensure it's enabled right away
             addSkel["access"] = ["root"]
             addSkel["password"] = pw