+++ core/skeleton.py
@@ -255,7 +255,10 @@ class SkeletonInstance:
         self.renderAccessedValues = {}
 
     def structure(self) -> dict:
-        return {key: bone.structure() for key, bone in self.items()}
+        return {
+            key: bone.structure() | {"sortindex": i}
+            for i, (key, bone) in enumerate(self.items())
+        }
 
     def __deepcopy__(self, memodict):
         res = self.clone()
@@ -1283,7 +1286,13 @@ class SkelList(list):
         :vartype cursor: str
     """
 
-    __slots__ = ["baseSkel", "getCursor", "customQueryInfo", "renderPreparation"]
+    __slots__ = (
+        "baseSkel",
+        "customQueryInfo",
+        "getCursor",
+        "get_orders",
+        "renderPreparation",
+    )
 
     def __init__(self, baseSkel=None):
         """
@@ -1292,6 +1301,7 @@ class SkelList(list):
         super(SkelList, self).__init__()
         self.baseSkel = baseSkel or {}
         self.getCursor = lambda: None
+        self.get_orders = lambda: None
         self.renderPreparation = None
         self.customQueryInfo = {}